[RISC-V] Final fixes

+++ src/coreclr/jit/emitriscv64.cpp
@@ -2889,7 +2889,7 @@ BYTE* emitter::emitOutputInstr_OptsI8(BYTE* dst, const instrDesc* id, ssize_t im
     {
         // special for INT64_MAX or UINT32_MAX
         dst += emitOutput_ITypeInstr(dst, INS_addi, reg1, REG_R0, NBitMask<12>);
-        const ssize_t shiftValue = (immediate == INT64_MAX) ? 1 : 32;
+        const unsigned shiftValue = (immediate == INT64_MAX) ? 1 : 32;
         dst += emitOutput_ITypeInstr(dst, INS_srli, reg1, reg1, shiftValue);
     }
     else
@@ -2964,7 +2964,7 @@ BYTE* emitter::emitOutputInstr_OptsRcNoReloc(BYTE* dst, instruction* ins, unsign
     const regNumber rsvdReg = codeGen->rsGetRsvdReg();
 
     const instruction lastIns = (*ins == INS_jal) ? (*ins = INS_addi) : *ins;
-    const UINT32      high = immediate >> 11;
+    const ssize_t     high = immediate >> 11;
 
     dst += emitOutput_UTypeInstr(dst, INS_lui, rsvdReg, UpperNBitsOfWordSignExtend<20>(high));
     dst += emitOutput_ITypeInstr(dst, INS_addi, rsvdReg, rsvdReg, LowerNBitsOfWord<12>(high));
